feat(llm-runtime): provide examples for transformers, vllm, and mlc-llm
