--- llama-3-2-3b-instruct/v0.1.0-vllm-cpu/instill.yaml
@@ -1,5 +1,0 @@
-build:
-  gpu: false
-  python_version: "3.11" # support only 3.11
-  python_packages:
-    - vllm==0.8.3
