+++ qwen-2-5-coder-0.5B-instruct/mlc-llm-cpu/instill.yaml
@@ -1,0 +1,3 @@
+build:
+  gpu: false
+  llm_runtime: "mlc-llm"
