+++ qwen-2-5-coder-0.5B-instruct/transformers-cpu/instill.yaml
@@ -1,0 +1,3 @@
+build:
+  gpu: false
+  llm_runtime: "transformers"
