--- qwen-2-5-coder-32b-instruct/v0.1.0/.gitignore
@@ -1,1 +1,0 @@
-*.gguf
