--- qwen-2-5-coder-32b-instruct/v0.1.0/instill.yaml
@@ -1,6 +1,0 @@
-build:
-  gpu: true
-  python_version: "3.11" # support only 3.11
-  cuda_version: "12.1"
-  python_packages:
-    - vllm==0.8.3
