--- qwen-2-5-coder-32b-instruct/v0.1.1/.gitignore
@@ -1,1 +1,0 @@
-Qwen2.5-Coder-32B-Instruct-AWQ
